type

+++ crates/task-impls/src/vid.rs
@@ -75,7 +75,7 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>, A: ConsensusApi<TYPES, I> + 
                 // calculate vid shares
                 let vid_disperse = spawn_blocking(move || {
                     #[allow(clippy::panic)]
-                    vid_scheme(num_storage_nodes as u32).disperse(&encoded_transactions).unwrap_or_else(|err|panic!("VID disperse failure:\n\t(num_storage nodes,payload_byte_len)=({num_storage_nodes},{})\n\terror: : {err}", encoded_transactions.len()))
+                    vid_scheme(num_storage_nodes).disperse(&encoded_transactions).unwrap_or_else(|err|panic!("VID disperse failure:\n\t(num_storage nodes,payload_byte_len)=({num_storage_nodes},{})\n\terror: : {err}", encoded_transactions.len()))
                 })
                 .await;
 
@@ -160,7 +160,7 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>, A: ConsensusApi<TYPES, I> + 
                 let num_nodes = self.membership.total_nodes();
                 let vid_disperse = spawn_blocking(move || {
                     #[allow(clippy::panic)]
-                    vid_scheme(num_nodes as u32).disperse(&txns).unwrap_or_else(|err|panic!("VID disperse failure:\n\t(num_storage nodes,payload_byte_len)=({num_nodes},{})\n\terror: : {err}", txns.len()))
+                    vid_scheme(num_nodes).disperse(&txns).unwrap_or_else(|err|panic!("VID disperse failure:\n\t(num_storage nodes,payload_byte_len)=({num_nodes},{})\n\terror: : {err}", txns.len()))
                 })
                 .await;
                 #[cfg(async_executor_impl = "tokio")]
